requirements.txt
ollama # Used to call the Ollama API for LLMs
requests # Used for making HTTP requests to APIs
pyyaml # Used for reading and writing YAML files
pyttsx3 # Text-to-speech conversion library
SpeechRecognition # Speech recognition library
pyaudio # Required for audio input/output in speech recognition
beautifulsoup4 # For parsing HTML and XML documents

# The following libraries are for agent management and communication
smolagents
smolagents[tool]
smolagents[litellm]
llama-index
duckduckgo_search
serpapi
google-search-results

# The following libraries are for better speech communication, but not working yet due to api issues
#elevenlabs==0.2.26
#bark